+++ spec/bali/components/delete_link_spec.rb
@@ -32,7 +32,7 @@ RSpec.describe Bali::DeleteLink::Component, type: :component do
   end
 
   it 'add a css class to the link' do
-    @options.merge!(classes: 'is-large')
+    @options.merge!(class: 'is-large')
     render_inline(component)
 
     expect(rendered_component).to have_css 'button.has-text-danger.is-text.is-large'
@@ -42,6 +42,14 @@ RSpec.describe Bali::DeleteLink::Component, type: :component do
     @options.merge!(href: nil)
 
     expect { render_inline(component) }.to raise_error(Bali::DeleteLink::Component::MissingURL)
+  end
+
+  it 'renders a delete link with custom form classes' do
+    @options.merge!(form_class: 'has-background-success')
+    render_inline(component)
+
+    expect(rendered_component).to have_css 'button.has-text-danger.is-text', text: 'Delete'
+    expect(rendered_component).to have_css 'form.button_to.has-background-success'
   end
 
   context 'with active record model' do