+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add `Avatar::Component`. With this component we'll be able to see a preview of the image we want as an avatar.
 
+## [0.24.4] - 2022-07-05
+
+- Let `DeleteLink::Component` receive form classes for the `buttton_to` tag.
+
 ## [0.24.3] - 2022-07-05
 
 - Add `type="button"` to Carousel controls (`arrows`, `bullets`).
@@ -35,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.23.2] - 2022-07-04
 
-- Set `Delete` as default name for `Deletelink::Component`
+- Set `Delete` as default name for `DeleteLink::Component`
 
 ## [0.23.1] - 2022-07-04